Fix duplicate file names unique prefix bug
File names with the unique prefixes, such as `a/file.txt` and
`b/file.txt`, in the same folder should be associated with their
respective EntryData in `filesZip`. Because the file paths need
to be unique to create such an association, `s3Files` needs to
preserve folder structure. Note that this only applies when
`filesZip` is present.